s/extend-webpack-config/enhance-webpack-config
