+++ packages/react-cosmos/src/server/web/webpack/__tests__/enhance-webpack-config/hmr.js
@@ -1,4 +1,4 @@
-import extendWebpackConfig from '../../extend-webpack-config';
+import enhanceWebpackConfig from '../../enhance-webpack-config';
 
 jest.mock('react-cosmos-config', () => ({
   hasUserCosmosConfig: () => true,
@@ -19,7 +19,7 @@ const webpack = {
 };
 
 const getConfig = () =>
-  extendWebpackConfig({
+  enhanceWebpackConfig({
     webpack,
     userWebpackConfig: {}
   });
